Refactor OAuth flow to use direct redirect and username selection page
